Added a good way for switching back to gui_dev (#49).

+++ setup_new.py
@@ -58,42 +58,44 @@ if __name__ == '__main__':
     # Ask user for the experiment name.
     experiment_name = input('Enter a name for your experiment: ')
 
-    # Check if an experiment directory with this name already exists.
-    if os.path.isdir(experiment_name):
-        # Ask for permission to remove the directory.
-        rm_dir = input('An experiment directory with this name already exists. '
-                       'Do you want to overwrite it? (yes/no): ')
-        if rm_dir.lower() in ('yes', 'y'):
-            # Remove and recreate if permission is given.
-            shutil.rmtree(experiment_name)
-            os.mkdir(experiment_name)
+    # Don't do any of the create and/or replace actions in case of gui_dev
+    if not experiment_name == 'gui_dev':
+        # Check if an experiment directory with this name already exists.
+        if os.path.isdir(experiment_name):
+            # Ask for permission to remove the directory.
+            rm_dir = input('An experiment directory with this name already exists. '
+                           'Do you want to overwrite it? (yes/no): ')
+            if rm_dir.lower() in ('yes', 'y'):
+                # Remove and recreate if permission is given.
+                shutil.rmtree(experiment_name)
+                os.mkdir(experiment_name)
+            else:
+                # Otherwise, do nothing.
+                print(f'Not overwriting the experiment directory {experiment_name}.')
         else:
-            # Otherwise, do nothing.
-            print(f'Not overwriting the experiment directory {experiment_name}.')
-    else:
-        # Create a new directory if it doesn't exist yet.
-        os.mkdir(experiment_name)
+            # Create a new directory if it doesn't exist yet.
+            os.mkdir(experiment_name)
 
-    # Check if an input file already exists for this experiment name.
-    if os.path.isfile(f'{experiment_name}.palila'):
-        # Ask for permission to remove the input file.
-        rm_file = input('An experiment input file with this name already exists. '
-                        'Do you want to overwrite it? (yes/no): ')
-        if rm_file.lower() in ('yes', 'y'):
-            # Remove and recreate if permission is given.
-            os.remove(f'{experiment_name}.palila')
+        # Check if an input file already exists for this experiment name.
+        if os.path.isfile(f'{experiment_name}.palila'):
+            # Ask for permission to remove the input file.
+            rm_file = input('An experiment input file with this name already exists. '
+                            'Do you want to overwrite it? (yes/no): ')
+            if rm_file.lower() in ('yes', 'y'):
+                # Remove and recreate if permission is given.
+                os.remove(f'{experiment_name}.palila')
+                with open(f'{experiment_name}.palila', 'w') as palila_file:
+                    palila_file.writelines(NEW_STRING)
+            else:
+                # Otherwise, do nothing.
+                print(f'Not overwriting the experiment input file {experiment_name}.palila.')
+        else:
+            # Create a new input file if it doesn't exist yet.
             with open(f'{experiment_name}.palila', 'w') as palila_file:
                 palila_file.writelines(NEW_STRING)
-        else:
-            # Otherwise, do nothing.
-            print(f'Not overwriting the experiment input file {experiment_name}.palila.')
-    else:
-        # Create a new input file if it doesn't exist yet.
-        with open(f'{experiment_name}.palila', 'w') as palila_file:
-            palila_file.writelines(NEW_STRING)
 
-    # Copy the tone file to the experiment directory
-    shutil.copyfile('gui_dev/tone500Hz.wav', f'{experiment_name}/tone500Hz.wav')
+        # Copy the tone file to the experiment directory
+        shutil.copyfile('gui_dev/tone500Hz.wav', f'{experiment_name}/tone500Hz.wav')
 
     # Read the current main python file of the GUI
     with open('main.py', 'r') as py_file:
